#121: Cleanup unused platform support from libS2 (#122)

+++ components/zwave/zwave_transports/s2/libs/zw-libs2/protocol/test/test_span.c
@@ -199,7 +199,7 @@ void S2_save_nls_state(void)
 /**
  * @brief Stub for clock_time - timer tick
  */
-uint32_t clock_time(void)
+unsigned long clock_time(void)
 {
     return 0;
 }
